+++ spec/eve_online/esi/base_spec.rb
@@ -7,36 +7,26 @@ describe EveOnline::ESI::Base do
 
   describe "#initialize" do
     context "with options" do
-      let(:parser) { double }
-
       let(:options) do
         {
           token: "token123",
-          parser: parser,
           read_timeout: 30,
           open_timeout: 45,
+          write_timeout: 50,
           etag: "6f2d3caa79a79bc9e61aa058e18905faac5e293fa1729637648ce9a1",
           language: "ru"
-        }.tap do |hash|
-          if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("2.6.0")
-            hash[:write_timeout] = 50
-          end
-        end
+        }
       end
 
       subject { described_class.new(options) }
 
       its(:token) { should eq("token123") }
 
-      its(:parser) { should eq(parser) }
-
       its(:_read_timeout) { should eq(30) }
 
       its(:_open_timeout) { should eq(45) }
 
-      if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("2.6.0")
-        its(:_write_timeout) { should eq(50) }
-      end
+      its(:_write_timeout) { should eq(50) }
 
       its(:_etag) { should eq("6f2d3caa79a79bc9e61aa058e18905faac5e293fa1729637648ce9a1") }
 
@@ -46,15 +36,11 @@ describe EveOnline::ESI::Base do
     context "without options" do
       its(:token) { should eq(nil) }
 
-      its(:parser) { should eq(JSON) }
-
       its(:_read_timeout) { should eq(60) }
 
       its(:_open_timeout) { should eq(60) }
 
-      if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("2.6.0")
-        its(:_write_timeout) { should eq(60) }
-      end
+      its(:_write_timeout) { should eq(60) }
 
       its(:_etag) { should eq(nil) }
 
@@ -86,17 +72,21 @@ describe EveOnline::ESI::Base do
   end
 
   describe "#http_method" do
-    specify { expect(subject.http_method).to eq("Get") }
+    specify { expect(subject.http_method).to eq(:get) }
   end
 
   describe "#read_timeout" do
     before do
       #
-      # subject.client.read_timeout
+      # subject.connection.options.read_timeout
       #
-      expect(subject).to receive(:client) do
+      expect(subject).to receive(:connection) do
         double.tap do |a|
-          expect(a).to receive(:read_timeout)
+          expect(a).to receive(:options) do
+            double.tap do |b|
+              expect(b).to receive(:read_timeout)
+            end
+          end
         end
       end
     end
@@ -109,11 +99,15 @@ describe EveOnline::ESI::Base do
 
     before do
       #
-      # subject.client.read_timeout = value
+      # subject.connection.options.read_timeout = value
       #
-      expect(subject).to receive(:client) do
+      expect(subject).to receive(:connection) do
         double.tap do |a|
-          expect(a).to receive(:read_timeout=).with(value)
+          expect(a).to receive(:options) do
+            double.tap do |b|
+              expect(b).to receive(:read_timeout=).with(value)
+            end
+          end
         end
       end
     end
@@ -124,11 +118,15 @@ describe EveOnline::ESI::Base do
   describe "#open_timeout" do
     before do
       #
-      # subject.client.open_timeout
+      # subject.connection.options.open_timeout
       #
-      expect(subject).to receive(:client) do
+      expect(subject).to receive(:connection) do
         double.tap do |a|
-          expect(a).to receive(:open_timeout)
+          expect(a).to receive(:options) do
+            double.tap do |b|
+              expect(b).to receive(:open_timeout)
+            end
+          end
         end
       end
     end
@@ -141,11 +139,15 @@ describe EveOnline::ESI::Base do
 
     before do
       #
-      # subject.client.open_timeout = value
+      # subject.connection.options.open_timeout = value
       #
-      expect(subject).to receive(:client) do
+      expect(subject).to receive(:connection) do
         double.tap do |a|
-          expect(a).to receive(:open_timeout=).with(value)
+          expect(a).to receive(:options) do
+            double.tap do |b|
+              expect(b).to receive(:open_timeout=).with(value)
+            end
+          end
         end
       end
     end
@@ -153,38 +155,44 @@ describe EveOnline::ESI::Base do
     specify { expect { subject.send(:open_timeout=, value) }.not_to raise_error }
   end
 
-  if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("2.6.0")
-    describe "#write_timeout" do
-      before do
-        #
-        # subject.client.write_timeout
-        #
-        expect(subject).to receive(:client) do
-          double.tap do |a|
-            expect(a).to receive(:write_timeout)
+  describe "#write_timeout" do
+    before do
+      #
+      # subject.connection.options.write_timeout
+      #
+      expect(subject).to receive(:connection) do
+        double.tap do |a|
+          expect(a).to receive(:options) do
+            double.tap do |b|
+              expect(b).to receive(:write_timeout)
+            end
           end
         end
       end
-
-      specify { expect { subject.write_timeout }.not_to raise_error }
     end
 
-    describe "#write_timeout=" do
-      let(:value) { double }
+    specify { expect { subject.write_timeout }.not_to raise_error }
+  end
 
-      before do
-        #
-        # subject.client.write_timeout = value
-        #
-        expect(subject).to receive(:client) do
-          double.tap do |a|
-            expect(a).to receive(:write_timeout=).with(value)
+  describe "#write_timeout=" do
+    let(:value) { double }
+
+    before do
+      #
+      # subject.connection.options.write_timeout = value
+      #
+      expect(subject).to receive(:connection) do
+        double.tap do |a|
+          expect(a).to receive(:options) do
+            double.tap do |b|
+              expect(b).to receive(:write_timeout=).with(value)
+            end
           end
         end
       end
-
-      specify { expect { subject.send(:write_timeout=, value) }.not_to raise_error }
     end
+
+    specify { expect { subject.send(:write_timeout=, value) }.not_to raise_error }
   end
 
   describe "#etag=" do
@@ -194,15 +202,15 @@ describe EveOnline::ESI::Base do
   describe "#etag" do
     let(:resource) { double }
 
-    let(:header) { double }
+    let(:headers) { double }
 
     let(:etag) { double }
 
     before { expect(subject).to receive(:resource).and_return(resource) }
 
-    before { expect(resource).to receive(:header).and_return(header) }
+    before { expect(resource).to receive(:headers).and_return(headers) }
 
-    before { expect(header).to receive(:[]).with("Etag").and_return(etag) }
+    before { expect(headers).to receive(:[]).with("etag").and_return(etag) }
 
     before { expect(etag).to receive(:gsub).with("W/", "").and_return(etag) }
 
@@ -218,17 +226,17 @@ describe EveOnline::ESI::Base do
   describe "#total_pages" do
     let(:resource) { double }
 
-    let(:header) { double }
+    let(:headers) { double }
 
-    let(:pages) { double }
+    let(:value) { double }
 
     before { expect(subject).to receive(:resource).and_return(resource) }
 
-    before { expect(resource).to receive(:header).and_return(header) }
+    before { expect(resource).to receive(:headers).and_return(headers) }
 
-    before { expect(header).to receive(:[]).with("X-Pages").and_return(pages) }
+    before { expect(headers).to receive(:[]).with("x-pages").and_return(value) }
 
-    before { expect(pages).to receive(:to_i) }
+    before { expect(value).to receive(:to_i) }
 
     specify { expect { subject.total_pages }.not_to raise_error }
   end
@@ -236,17 +244,17 @@ describe EveOnline::ESI::Base do
   describe "#error_limit_remain" do
     let(:resource) { double }
 
-    let(:header) { double }
+    let(:headers) { double }
 
-    let(:pages) { double }
+    let(:value) { double }
 
     before { expect(subject).to receive(:resource).and_return(resource) }
 
-    before { expect(resource).to receive(:header).and_return(header) }
+    before { expect(resource).to receive(:headers).and_return(headers) }
 
-    before { expect(header).to receive(:[]).with("X-ESI-Error-Limit-Remain").and_return(pages) }
+    before { expect(headers).to receive(:[]).with("x-esi-error-limit-remain").and_return(value) }
 
-    before { expect(pages).to receive(:to_i) }
+    before { expect(value).to receive(:to_i) }
 
     specify { expect { subject.error_limit_remain }.not_to raise_error }
   end
@@ -254,164 +262,176 @@ describe EveOnline::ESI::Base do
   describe "#error_limit_reset" do
     let(:resource) { double }
 
-    let(:header) { double }
+    let(:headers) { double }
 
-    let(:pages) { double }
+    let(:value) { double }
 
     before { expect(subject).to receive(:resource).and_return(resource) }
 
-    before { expect(resource).to receive(:header).and_return(header) }
+    before { expect(resource).to receive(:headers).and_return(headers) }
 
-    before { expect(header).to receive(:[]).with("X-ESI-Error-Limit-Reset").and_return(pages) }
+    before { expect(headers).to receive(:[]).with("x-esi-error-limit-reset").and_return(value) }
 
-    before { expect(pages).to receive(:to_i) }
+    before { expect(value).to receive(:to_i) }
 
     specify { expect { subject.error_limit_reset }.not_to raise_error }
   end
 
+  describe "#request_id" do
+    let(:resource) { double }
+
+    let(:headers) { double }
+
+    before { expect(subject).to receive(:resource).and_return(resource) }
+
+    before { expect(resource).to receive(:headers).and_return(headers) }
+
+    before { expect(headers).to receive(:[]).with("x-esi-request-id") }
+
+    specify { expect { subject.request_id }.not_to raise_error }
+  end
+
+  describe "#connection" do
+    # TODO: write
+  end
+
   describe "#client" do
-    context "when @client set" do
-      let(:client) { double }
+    # context "when @client set" do
+    #   let(:client) { double }
+    #
+    #   before { subject.instance_variable_set(:@client, client) }
+    #
+    #   specify { expect(subject.client).to eq(client) }
+    # end
 
-      before { subject.instance_variable_set(:@client, client) }
-
-      specify { expect(subject.client).to eq(client) }
-    end
-
-    context "when @client not set" do
-      let(:host) { double }
-
-      let(:port) { double }
-
-      let(:uri) { double(host: host, port: port) }
-
-      let(:http) { instance_double(Net::HTTP) }
-
-      let(:_read_timeout) { double }
-
-      let(:_open_timeout) { double }
-
-      if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("2.6.0")
-        let(:_write_timeout) { double }
-      end
-
-      before { expect(subject).to receive(:uri).and_return(uri).twice }
-
-      before { expect(Net::HTTP).to receive(:new).with(host, port).and_return(http) }
-
-      before { expect(subject).to receive(:_read_timeout).and_return(_read_timeout) }
-
-      before { expect(subject).to receive(:_open_timeout).and_return(_open_timeout) }
-
-      if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("2.6.0")
-        before { expect(subject).to receive(:_write_timeout).and_return(_write_timeout) }
-      end
-
-      before { expect(http).to receive(:read_timeout=).with(_read_timeout) }
-
-      before { expect(http).to receive(:open_timeout=).with(_open_timeout) }
-
-      if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("2.6.0")
-        before { expect(http).to receive(:write_timeout=).with(_write_timeout) }
-      end
-
-      before { expect(http).to receive(:use_ssl=).with(true) }
-
-      before { expect(http).to receive(:verify_mode=).with(OpenSSL::SSL::VERIFY_PEER) }
-
-      specify { expect { subject.client }.not_to raise_error }
-
-      specify { expect { subject.client }.to change { subject.instance_variable_get(:@client) }.from(nil).to(http) }
-    end
+    # context "when @client not set" do
+    #   let(:host) { double }
+    #
+    #   let(:port) { double }
+    #
+    #   let(:uri) { double(host: host, port: port) }
+    #
+    #   let(:http) { instance_double(Net::HTTP) }
+    #
+    #   let(:_read_timeout) { double }
+    #
+    #   let(:_open_timeout) { double }
+    #
+    #   let(:_write_timeout) { double }
+    #
+    #   before { expect(subject).to receive(:uri).and_return(uri).twice }
+    #
+    #   before { expect(Net::HTTP).to receive(:new).with(host, port).and_return(http) }
+    #
+    #   before { expect(subject).to receive(:_read_timeout).and_return(_read_timeout) }
+    #
+    #   before { expect(subject).to receive(:_open_timeout).and_return(_open_timeout) }
+    #
+    #   before { expect(subject).to receive(:_write_timeout).and_return(_write_timeout) }
+    #
+    #   before { expect(http).to receive(:read_timeout=).with(_read_timeout) }
+    #
+    #   before { expect(http).to receive(:open_timeout=).with(_open_timeout) }
+    #
+    #   before { expect(http).to receive(:write_timeout=).with(_write_timeout) }
+    #
+    #   before { expect(http).to receive(:use_ssl=).with(true) }
+    #
+    #   before { expect(http).to receive(:verify_mode=).with(OpenSSL::SSL::VERIFY_PEER) }
+    #
+    #   specify { expect { subject.client }.not_to raise_error }
+    #
+    #   specify { expect { subject.client }.to change { subject.instance_variable_get(:@client) }.from(nil).to(http) }
+    # end
   end
 
-  describe "#request" do
-    context "when @request set" do
-      let(:request) { double }
-
-      before { subject.instance_variable_set(:@request, request) }
-
-      specify { expect(subject.request).to eq(request) }
-    end
-
-    context "when @request not set" do
-      let(:request_uri) { double }
-
-      let(:uri) { double(request_uri: request_uri) }
-
-      let(:user_agent) { double }
-
-      before { expect(subject).to receive(:uri).and_return(uri) }
-
-      before { expect(subject).to receive(:user_agent).and_return(user_agent) }
-
-      before { expect(subject).to receive(:http_method).and_return(http_method).exactly(3).times }
-
-      before { expect(request).to receive(:[]=).with("User-Agent", user_agent).and_return(request) }
-
-      before { expect(request).to receive(:[]=).with("Accept", "application/json").and_return(request) }
-
-      context "when http method Get and without token and etag" do
-        let(:http_method) { "Get" }
-
-        let(:request) { instance_double(Net::HTTP::Get) }
-
-        before { expect(Net::HTTP::Get).to receive(:new).with(request_uri).and_return(request) }
-
-        before { expect(subject).to receive(:token).and_return(nil) }
-
-        before { expect(subject).to receive(:_etag).and_return(nil) }
-
-        specify { expect { subject.request }.not_to raise_error }
-
-        specify { expect { subject.request }.to change { subject.instance_variable_get(:@request) }.from(nil).to(request) }
-      end
-
-      context "when http method Get and with token and etag" do
-        let(:http_method) { "Get" }
-
-        let(:request) { instance_double(Net::HTTP::Get) }
-
-        before { expect(Net::HTTP::Get).to receive(:new).with(request_uri).and_return(request) }
-
-        let(:token) { "token123" }
-
-        let(:_etag) { "etag" }
-
-        before { expect(subject).to receive(:token).and_return(token).twice }
-
-        before { expect(subject).to receive(:_etag).and_return(_etag).twice }
-
-        before { expect(request).to receive(:[]=).with("Authorization", "Bearer token123").and_return(request) }
-
-        before { expect(request).to receive(:[]=).with("If-None-Match", "etag").and_return(request) }
-
-        specify { expect { subject.request }.not_to raise_error }
-
-        specify { expect { subject.request }.to change { subject.instance_variable_get(:@request) }.from(nil).to(request) }
-      end
-
-      context "when http method Post" do
-        let(:http_method) { "Post" }
-
-        let(:request) { instance_double(Net::HTTP::Post) }
-
-        before { expect(Net::HTTP::Post).to receive(:new).with(request_uri).and_return(request) }
-
-        let(:payload) { double }
-
-        before { expect(subject).to receive(:payload).and_return(payload) }
-
-        before { expect(request).to receive(:[]=).with("Content-Type", "application/json").and_return(request) }
-
-        before { expect(request).to receive(:body=).with(payload) }
-
-        specify { expect { subject.request }.not_to raise_error }
-
-        specify { expect { subject.request }.to change { subject.instance_variable_get(:@request) }.from(nil).to(request) }
-      end
-    end
-  end
+  # describe "#request" do
+  #   context "when @request set" do
+  #     let(:request) { double }
+  #
+  #     before { subject.instance_variable_set(:@request, request) }
+  #
+  #     specify { expect(subject.request).to eq(request) }
+  #   end
+  #
+  #   context "when @request not set" do
+  #     let(:request_uri) { double }
+  #
+  #     let(:uri) { double(request_uri: request_uri) }
+  #
+  #     let(:user_agent) { double }
+  #
+  #     before { expect(subject).to receive(:uri).and_return(uri) }
+  #
+  #     before { expect(subject).to receive(:user_agent).and_return(user_agent) }
+  #
+  #     before { expect(subject).to receive(:http_method).and_return(http_method).exactly(3).times }
+  #
+  #     before { expect(request).to receive(:[]=).with("User-Agent", user_agent).and_return(request) }
+  #
+  #     before { expect(request).to receive(:[]=).with("Accept", "application/json").and_return(request) }
+  #
+  #     context "when http method Get and without token and etag" do
+  #       let(:http_method) { "Get" }
+  #
+  #       let(:request) { instance_double(Net::HTTP::Get) }
+  #
+  #       before { expect(Net::HTTP::Get).to receive(:new).with(request_uri).and_return(request) }
+  #
+  #       before { expect(subject).to receive(:token).and_return(nil) }
+  #
+  #       before { expect(subject).to receive(:_etag).and_return(nil) }
+  #
+  #       specify { expect { subject.request }.not_to raise_error }
+  #
+  #       specify { expect { subject.request }.to change { subject.instance_variable_get(:@request) }.from(nil).to(request) }
+  #     end
+  #
+  #     context "when http method Get and with token and etag" do
+  #       let(:http_method) { "Get" }
+  #
+  #       let(:request) { instance_double(Net::HTTP::Get) }
+  #
+  #       before { expect(Net::HTTP::Get).to receive(:new).with(request_uri).and_return(request) }
+  #
+  #       let(:token) { "token123" }
+  #
+  #       let(:_etag) { "etag" }
+  #
+  #       before { expect(subject).to receive(:token).and_return(token).twice }
+  #
+  #       before { expect(subject).to receive(:_etag).and_return(_etag).twice }
+  #
+  #       before { expect(request).to receive(:[]=).with("Authorization", "Bearer token123").and_return(request) }
+  #
+  #       before { expect(request).to receive(:[]=).with("If-None-Match", "etag").and_return(request) }
+  #
+  #       specify { expect { subject.request }.not_to raise_error }
+  #
+  #       specify { expect { subject.request }.to change { subject.instance_variable_get(:@request) }.from(nil).to(request) }
+  #     end
+  #
+  #     context "when http method Post" do
+  #       let(:http_method) { "Post" }
+  #
+  #       let(:request) { instance_double(Net::HTTP::Post) }
+  #
+  #       before { expect(Net::HTTP::Post).to receive(:new).with(request_uri).and_return(request) }
+  #
+  #       let(:payload) { double }
+  #
+  #       before { expect(subject).to receive(:payload).and_return(payload) }
+  #
+  #       before { expect(request).to receive(:[]=).with("Content-Type", "application/json").and_return(request) }
+  #
+  #       before { expect(request).to receive(:body=).with(payload) }
+  #
+  #       specify { expect { subject.request }.not_to raise_error }
+  #
+  #       specify { expect { subject.request }.to change { subject.instance_variable_get(:@request) }.from(nil).to(request) }
+  #     end
+  #   end
+  # end
 
   describe "#uri" do
     context "when @uri set" do
@@ -459,8 +479,7 @@ describe EveOnline::ESI::Base do
 
         before do
           #
-          # URI::HTTPS.build(host: API_HOST,
-          #                  path: path) # => uri
+          # URI::HTTPS.build(host: API_HOST, path: path) # => uri
           #
           expect(URI::HTTPS).to receive(:build).with(host: described_class::API_HOST,
                                                      path: path).and_return(uri)
@@ -516,41 +535,41 @@ describe EveOnline::ESI::Base do
       specify { expect(subject.resource).to eq(resource) }
     end
 
-    context "when @resource not set" do
-      let(:resource) { double }
+    # context "when @resource not set" do
+    #   let(:resource) { double }
+    #
+    #   let(:client) { double }
+    #
+    #   let(:request) { double }
+    #
+    #   before { expect(subject).to receive(:request).and_return(request) }
+    #
+    #   before { expect(subject).to receive(:client).and_return(client) }
+    #
+    #   before { expect(client).to receive(:request).with(request).and_return(resource) }
+    #
+    #   specify { expect { subject.resource }.not_to raise_error }
+    #
+    #   specify { expect { subject.resource }.to change { subject.instance_variable_get(:@resource) }.from(nil).to(resource) }
+    # end
 
-      let(:client) { double }
-
-      let(:request) { double }
-
-      before { expect(subject).to receive(:request).and_return(request) }
-
-      before { expect(subject).to receive(:client).and_return(client) }
-
-      before { expect(client).to receive(:request).with(request).and_return(resource) }
-
-      specify { expect { subject.resource }.not_to raise_error }
-
-      specify { expect { subject.resource }.to change { subject.instance_variable_get(:@resource) }.from(nil).to(resource) }
-    end
-
-    context "when Net::HTTP throw Net::OpenTimeout" do
-      before { expect(subject).to receive(:client).and_raise(Net::OpenTimeout) }
-
-      specify { expect { subject.resource }.to raise_error(EveOnline::Exceptions::Timeout) }
-    end
-
-    context "when Net::HTTP throw Net::ReadTimeout" do
-      before { expect(subject).to receive(:client).and_raise(Net::ReadTimeout) }
+    context "when throw Faraday::ConnectionFailed" do
+      before { expect(subject).to receive(:connection).and_raise(Faraday::ConnectionFailed) }
 
       specify { expect { subject.resource }.to raise_error(EveOnline::Exceptions::Timeout) }
     end
 
-    context "when Net::HTTP throw Net::WriteTimeout" do
-      before { expect(subject).to receive(:client).and_raise(Net::WriteTimeout) }
+    context "when Faraday::TimeoutError" do
+      before { expect(subject).to receive(:connection).and_raise(Faraday::TimeoutError) }
 
       specify { expect { subject.resource }.to raise_error(EveOnline::Exceptions::Timeout) }
     end
+
+    # context "when Net::HTTP throw Net::WriteTimeout" do
+    #   before { expect(subject).to receive(:client).and_raise(Net::WriteTimeout) }
+    #
+    #   specify { expect { subject.resource }.to raise_error(EveOnline::Exceptions::Timeout) }
+    # end
   end
 
   describe "#not_modified?" do
@@ -697,15 +716,9 @@ describe EveOnline::ESI::Base do
     end
 
     context "when @response not set" do
-      let(:parser) { double }
-
       let(:content) { "some content to parse" }
 
       before { expect(subject).to receive(:content).and_return(content) }
-
-      before { expect(subject).to receive(:parser).and_return(parser) }
-
-      before { expect(parser).to receive(:parse).with(content) }
 
       specify { expect { subject.response }.not_to raise_error }
     end
@@ -718,7 +731,7 @@ describe EveOnline::ESI::Base do
 
     before do
       #
-      # ActiveSupport::TimeZone['UTC'].parse(value)
+      # ActiveSupport::TimeZone["UTC"].parse(value)
       #
       expect(ActiveSupport::TimeZone).to receive(:[]).with("UTC") do
         double.tap do |a|